requirements.txt
numpy
Pillow
cryptography
kivy